Rename, use next for iteration.
